requirements.txt
aiofiles==0.8.0
aioftp==0.21.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
APScheduler==3.9.1
asgiref==3.5.0
astroid==2.9.3
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.1.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fastapi==0.75.0
fastapi-utils==0.2.1
firebase-admin==5.2.0
frozenlist==1.3.0
google-api-core==2.7.1
google-api-python-client==2.40.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
loguru==0.6.0
mccabe==0.6.1
motor==2.5.1
msgpack==1.0.3
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.9.0
pylint==2.12.2
pymongo==3.12.3
pyparsing==3.0.7
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
wrapt==1.13.3
yarl==1.7.2
